+++ Changelog.rst
@@ -5,6 +5,7 @@ ChangeLog
 ----------------
 
 - Allow to search for more than one kind of bar code at once
+- **deprecate** scan_codes(str, Image) in favor of scan_codes(list, Image)
 
 
 2.0 (2018-01-24)